feat(backend): refactored data serializers
made a common base class for the DataLayer & MapRequest ViewSerializers.
Added a geometry_wkt field to the MapRequestSerializer.
