+++ server/safers/data/serializers/__init__.py
@@ -1,2 +1,2 @@
-from .serializers_datalayers import DataLayerSerializer
-from .serializers_maprequest import MapRequestSerializer
+from .serializers_datalayers import DataLayerViewSerializer
+from .serializers_maprequest import MapRequestSerializer, MapRequestViewSerializer
